TS-26565 Extracted TeamscaleMockServer and gradle dependencies to new module
