+++ common-system-test/build.gradle
@@ -1,0 +1,14 @@
+plugins {
+	id 'com.teamscale.java-convention'
+	id 'java-library'
+	// we do not enable code coverage recording for the system tests as we already need our agent attached
+	// it would conflict with JaCoCo's
+}
+
+dependencies {
+	api project(':teamscale-client')
+	api project(':tia-client')
+
+	api 'com.sparkjava:spark-core:2.9.3'
+	api 'com.squareup.moshi:moshi:1.12.0'
+}
