+++ settings.gradle
@@ -17,6 +17,7 @@ include ':teamscale-client'
 include ':sample-app'
 include ':impacted-test-engine'
 include ':tia-client'
+include ':common-system-test'
 include ':system-tests:tia-client-test'
 include ':system-tests:default-excludes-test'
 include ':system-tests:http-redirect-test'
