+++ system-tests/http-redirect-test/build.gradle
@@ -18,12 +18,5 @@ test {
 tasks.test.dependsOn(":agent:shadowJar")
 
 dependencies {
-	testImplementation project(':teamscale-client')
-
-	testImplementation 'com.sparkjava:spark-core:2.9.3'
-	testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.2'
-	testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.2'
-	testImplementation 'org.assertj:assertj-core:3.19.0'
-	testImplementation 'com.squareup.moshi:moshi:1.12.0'
-	testImplementation 'org.slf4j:slf4j-simple:1.7.9'
+	testImplementation project(':common-system-test')
 }
