+++ system-tests/http-redirect-test/src/test/java/com/teamscale/client/HttpRedirectSystemTest.java
@@ -1,5 +1,6 @@
 package com.teamscale.client;
 
+import com.teamscale.test.commons.TeamscaleMockServer;
 import org.junit.jupiter.api.Test;
 import retrofit2.Call;
 import retrofit2.Retrofit;
