+++ system-tests/tia-client-test/build.gradle
@@ -1,28 +1,22 @@
 plugins {
-    id 'com.teamscale.java-convention'
-    // we do not enable code coverage recording for the system tests as we already need our agent attached
-    // it would conflict with JaCoCo's
+	id 'com.teamscale.java-convention'
+	// we do not enable code coverage recording for the system tests as we already need our agent attached
+	// it would conflict with JaCoCo's
 }
 
 test {
-    /** These ports must match what is configured in the SystemTest class. */
-    def fakeTeamscalePort = 65432
-    def agentPort = 65433
-    def agentJar = project(':agent').tasks.shadowJar.outputs.files.files.first()
-    jvmArgs "-javaagent:$agentJar=http-server-port=$agentPort,tia-mode=teamscale-upload," +
-            "teamscale-server-url=http://localhost:$fakeTeamscalePort,teamscale-user=fake,teamscale-access-token=fake," +
-            "teamscale-project=p,teamscale-partition=part,mode=testwise,teamscale-commit=master:12345," +
-            "includes=*systemundertest.*"
+	/** These ports must match what is configured in the SystemTest class. */
+	def fakeTeamscalePort = 65432
+	def agentPort = 65433
+	def agentJar = project(':agent').tasks.shadowJar.outputs.files.files.first()
+	jvmArgs "-javaagent:$agentJar=http-server-port=$agentPort,tia-mode=teamscale-upload," +
+			"teamscale-server-url=http://localhost:$fakeTeamscalePort,teamscale-user=fake,teamscale-access-token=fake," +
+			"teamscale-project=p,teamscale-partition=part,mode=testwise,teamscale-commit=master:12345," +
+			"includes=*systemundertest.*"
 }
 
 dependencies {
-    implementation project(':tia-client')
-
-    testImplementation 'com.sparkjava:spark-core:2.9.3'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.2'
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.2'
-    testImplementation 'org.assertj:assertj-core:3.19.0'
-    testImplementation 'com.squareup.moshi:moshi:1.12.0'
+	implementation project(':common-system-test')
 }
 
 tasks.test.dependsOn(project(':agent').tasks.shadowJar)
